+++ src/index.js
@@ -2,7 +2,9 @@ import Diagram from './components/Diagram';
 
 export { default as Canvas } from './components/Canvas';
 export { default as Diagram } from './components/Diagram';
+export { default as CanvasControls } from './components/CanvasControls';
 export { default as useSchema } from './hooks/useSchema';
+export { default as useCanvasState } from './hooks/useCanvasState';
 export { default as createSchema } from './shared/functions/createSchema';
 export { validateNode } from './shared/functions/validators';
 export { validateNodes } from './shared/functions/validators';